refactor(frontend): move location of item
