+++ apps/frontend/app/components/media.tsx
@@ -2,16 +2,10 @@ import {
 	ActionIcon,
 	Anchor,
 	Avatar,
-	Box,
-	Center,
-	Flex,
 	Group,
-	Image,
 	Loader,
-	type MantineStyleProp,
 	Menu,
 	ScrollArea,
-	Skeleton,
 	Text,
 	ThemeIcon,
 	Tooltip,
@@ -22,14 +16,13 @@ import { useInViewport } from "@mantine/hooks";
 import { Form, Link } from "@remix-run/react";
 import {
 	EntityLot,
-	GridPacking,
 	MetadataGroupDetailsDocument,
 	PersonDetailsDocument,
 	SeenState,
 	UserReviewScale,
 	UserToMediaReason,
 } from "@ryot/generated/graphql/backend/graphql";
-import { changeCase, getInitials, isString, snakeCase } from "@ryot/ts-utils";
+import { changeCase, snakeCase } from "@ryot/ts-utils";
 import {
 	IconBackpack,
 	IconBookmarks,
@@ -38,10 +31,11 @@ import {
 	IconStarFilled,
 } from "@tabler/icons-react";
 import { useQuery } from "@tanstack/react-query";
-import type { ReactNode, Ref } from "react";
+import type { ReactNode } from "react";
 import { match } from "ts-pattern";
 import { withQuery } from "ufo";
 import {
+	BaseMediaDisplayItem,
 	DisplayThreePointReview,
 	MEDIA_DETAILS_HEIGHT,
 } from "~/components/common";
@@ -54,7 +48,6 @@ import {
 } from "~/lib/generals";
 import {
 	useConfirmSubmit,
-	useFallbackImageUrl,
 	useUserDetails,
 	useUserMetadataDetails,
 	useUserPreferences,
@@ -107,137 +100,6 @@ export const MediaScrollArea = (props: { children: ReactNode }) => {
 		<ScrollArea.Autosize mah={MEDIA_DETAILS_HEIGHT}>
 			{props.children}
 		</ScrollArea.Autosize>
-	);
-};
-
-const blackBgStyles = {
-	backgroundColor: "rgba(0, 0, 0, 0.75)",
-	borderRadius: 3,
-	padding: 2,
-} satisfies MantineStyleProp;
-
-export const BaseMediaDisplayItem = (props: {
-	isLoading: boolean;
-	name?: string;
-	altName?: string;
-	imageUrl?: string | null;
-	imageOverlay?: {
-		topRight?: ReactNode;
-		topLeft?: ReactNode;
-		bottomRight?: ReactNode;
-		bottomLeft?: ReactNode;
-	};
-	labels?: { right?: ReactNode; left?: ReactNode };
-	onImageClickBehavior: string | (() => Promise<void>);
-	nameRight?: ReactNode;
-	innerRef?: Ref<HTMLDivElement>;
-}) => {
-	const userPreferences = useUserPreferences();
-	const gridPacking = userPreferences.general.gridPacking;
-	const SurroundingElement = (iProps: { children: ReactNode }) =>
-		isString(props.onImageClickBehavior) ? (
-			<Anchor component={Link} to={props.onImageClickBehavior}>
-				{iProps.children}
-			</Anchor>
-		) : (
-			<Box onClick={props.onImageClickBehavior}>{iProps.children}</Box>
-		);
-	const defaultOverlayProps = {
-		style: { zIndex: 10, ...blackBgStyles },
-		pos: "absolute",
-	} as const;
-
-	return (
-		<Flex justify="space-between" direction="column" ref={props.innerRef}>
-			<Box pos="relative" w="100%">
-				<SurroundingElement>
-					<Tooltip label={props.name} position="top">
-						<Image
-							src={props.imageUrl}
-							radius="md"
-							style={{
-								cursor: "pointer",
-								...match(gridPacking)
-									.with(GridPacking.Normal, () => ({ height: 260 }))
-									.with(GridPacking.Dense, () => ({ height: 180 }))
-									.exhaustive(),
-							}}
-							alt={`Image for ${props.name}`}
-							className={classes.mediaImage}
-							styles={{
-								root: {
-									transitionProperty: "transform",
-									transitionTimingFunction: "cubic-bezier(0.4, 0, 0.2, 1)",
-									transitionDuration: "150ms",
-								},
-							}}
-							fallbackSrc={useFallbackImageUrl(
-								props.isLoading
-									? "Loading..."
-									: props.name
-										? getInitials(props.name)
-										: undefined,
-							)}
-						/>
-					</Tooltip>
-				</SurroundingElement>
-				{props.imageOverlay?.topLeft ? (
-					<Center top={5} left={5} {...defaultOverlayProps}>
-						{props.imageOverlay.topLeft}
-					</Center>
-				) : null}
-				{props.imageOverlay?.topRight ? (
-					<Center top={5} right={5} {...defaultOverlayProps}>
-						{props.imageOverlay.topRight}
-					</Center>
-				) : null}
-				{props.imageOverlay?.bottomLeft ? (
-					<Center bottom={5} left={5} {...defaultOverlayProps}>
-						{props.imageOverlay.bottomLeft}
-					</Center>
-				) : null}
-				{props.imageOverlay?.bottomRight ? (
-					<Center bottom={5} right={5} {...defaultOverlayProps}>
-						{props.imageOverlay.bottomRight}
-					</Center>
-				) : null}
-			</Box>
-			{props.isLoading ? (
-				<>
-					<Skeleton height={22} mt={10} />
-					<Skeleton height={22} mt={8} />
-				</>
-			) : (
-				<Flex
-					w="100%"
-					direction="column"
-					mt={2}
-					px={match(gridPacking)
-						.with(GridPacking.Normal, () => ({ base: 6, md: 3 }))
-						.with(GridPacking.Dense, () => ({ md: 2 }))
-						.exhaustive()}
-				>
-					<Flex w="100%" direction="row" justify="space-between">
-						<Text
-							size="sm"
-							c="dimmed"
-							visibleFrom={gridPacking === GridPacking.Dense ? "md" : undefined}
-						>
-							{props.labels?.left}
-						</Text>
-						<Text c="dimmed" size="sm">
-							{props.labels?.right}
-						</Text>
-					</Flex>
-					<Flex justify="space-between" align="center" mb="xs">
-						<Text w="100%" truncate fw="bold">
-							{props.altName ?? props.name}
-						</Text>
-						{props.nameRight}
-					</Flex>
-				</Flex>
-			)}
-		</Flex>
 	);
 };
